kafka_manager/pyproject.toml
[project]
name = "kafka-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka",
    "python-dotenv",
    "requests>=2.32.4",
]
